change text_ops_test
