+++ delta/data/task/text_cls_task.py
@@ -23,7 +23,6 @@ from delta.data.task.base_text_task import TextTask
 
 from delta.data.utils.common_utils import load_npy
 from delta.data.utils.common_utils import get_file_len
-from delta.data.utils.common_utils import load_cls_raw_data
 from delta.data.preprocess.text_ops import process_one_label_dataset
 from delta.data.preprocess.text_ops import load_dense_dataset
 from delta.data.preprocess.utils import load_vocab_dict
