+++ delta/data/task/text_seq_label_task.py
@@ -20,7 +20,6 @@ import tensorflow as tf
 from absl import logging
 
 from delta.data.task.base_text_task import TextTask
-from delta.data.utils.common_utils import load_seq_label_raw_data
 from delta.data.preprocess.text_ops import process_multi_label_dataset
 from delta.data.preprocess.utils import get_vocab_size
 from delta.utils.register import registers
