+++ delta/data/utils/common_utils.py
@@ -90,187 +90,6 @@ def clean_english_str(string):
   return string.strip().lower()
 
 
-#pylint: disable=too-many-locals
-def load_nlu_joint_raw_data(paths, mode, infer_no_label=False):
-  """Load raw data for sequence labeling"""
-  text = []
-  intent_label = []
-  slots_label = []
-  max_seq_len = 0
-  for path in paths:
-    success_count = 0
-    fail_count = 0
-    with open(path, 'r', encoding='utf-8') as file_input:
-      lines = file_input.readlines()
-      for i, line in enumerate(lines):
-        line = line.strip()
-        if mode == utils.INFER and infer_no_label:
-          text.append(line)
-        else:
-          sp = line.split('\t')
-          if len(sp) != 3:
-            logging.warning("Line no.{} not in standard format!".format(i))
-            fail_count += 1
-            continue
-          else:
-            success_count += 1
-          sent_len = len(sp[1].split(" "))
-          if sent_len > max_seq_len:
-            max_seq_len = sent_len
-          intent_label.append(sp[0])
-          slots_label.append(sp[1])
-          text.append(sp[2])
-
-      all_lines = len(lines)
-      del lines
-    logging.info("max_seq_len is {}".format(max_seq_len))
-    logging.info("Data loaded from {}. " \
-                 "Total {} lines, successfully load {} lines, " \
-                 "failed {} lines.".format(path, all_lines, success_count, fail_count))
-  if mode == utils.INFER and infer_no_label:
-    return text, ([], [])
-  return text, (intent_label, slots_label)
-
-
-#pylint: enable=too-many-locals
-
-
-# pylint: disable=too-many-branches
-def load_seq_label_raw_data(paths, mode, infer_no_label=False):
-  """Load raw data for sequence labeling"""
-  text = []
-  label = []
-  max_seq_len = 0
-  for path in paths:
-    success_count = 0
-    fail_count = 0
-    with open(path, 'r', encoding='utf-8') as file_input:
-      logging.info("Loading raw file from {}".format(path))
-      lines = file_input.readlines()
-      for i, line in enumerate(lines):
-        line = line.strip()
-        if mode == utils.INFER and infer_no_label:
-          text.append(line)
-        else:
-          sp = line.split('\t')
-          if len(sp) != 2:
-            logging.warning("Line no.{} not in standard format!".format(i))
-            fail_count += 1
-            continue
-          else:
-            success_count += 1
-          sent_len = len(sp[1].split(" "))
-          if sent_len > max_seq_len:
-            max_seq_len = sent_len
-          text.append(sp[1])
-          label.append(sp[0])
-      all_lines = len(lines)
-      del lines
-    logging.info("max_seq_len is {}".format(max_seq_len))
-    logging.info("Data loaded from {}. " \
-                 "Total {} lines, successfully load {} lines, " \
-                 "failed {} lines.".format(path, all_lines, success_count, fail_count))
-  if mode == utils.INFER and infer_no_label:
-    return text, []
-  return text, label
-
-
-def load_cls_raw_data(paths, mode, infer_no_label=False):
-  """Load raw data for classification."""
-  text = []
-  label = []
-  for path in paths:
-    logging.info("Loading raw file from {}".format(path))
-    success_count = 0
-    fail_count = 0
-    with open(path, 'r', encoding='utf-8') as file_input:
-      lines = file_input.readlines()
-      for i, line in enumerate(lines):
-        line = line.strip()
-        if mode == utils.INFER and infer_no_label:
-          text.append(line)
-        else:
-          sp = line.split('\t')
-          if len(sp) != 2:
-            logging.warning("Line no.{} not in standard format!".format(i))
-            fail_count += 1
-            continue
-          else:
-            success_count += 1
-          text.append(" ".join(sp[1:]))
-          label.append(sp[0])
-      all_lines = len(lines)
-      del lines
-    logging.info("Data loaded from {}. " \
-                 "Total {} lines, successfully load {} lines, " \
-                 "failed {} lines.".format(path, all_lines, success_count, fail_count))
-  if mode == utils.INFER and infer_no_label:
-    return text, []
-  return text, label
-
-
-# pylint: disable = too-many-locals
-def load_match_raw_data(paths, mode, infer_no_label=False):
-  """Load raw data for text match."""
-  text_right = []
-  text_left = []
-  label = []
-  for path in paths:
-    success_count = 0
-    fail_count = 0
-    with open(path, 'r', encoding='utf-8') as file_input:
-      logging.info("Loading raw file from {}".format(path))
-      lines = file_input.readlines()
-      for i, line in enumerate(lines):
-        line = line.strip()
-        sp = line.split('\t')
-        if mode == utils.INFER and infer_no_label:
-          if len(sp) != 2:
-            fail_count += 1
-            logging.warning("Line no.{} not in standard format!".format(i))
-            continue
-          else:
-            success_count += 1
-            text_left.append(sp[0])  # [sentence1, sentence2]
-            text_right.append(sp[1])
-        else:
-          if len(sp) != 3:
-            fail_count += 1
-            logging.warning("Line no.{} not in standard format!".format(i))
-            continue
-          else:
-            success_count += 1
-            text_left.append(sp[1])  # [label sentence1 sentence2]
-            text_right.append(sp[2])
-            label.append(sp[0])
-
-      all_lines = len(lines)
-      del lines
-    logging.info("Data loaded from {}. " \
-                 "Total {} lines, successfully load {} lines, " \
-                 "failed {} lines.".format(path, all_lines, success_count, fail_count))
-
-  if mode == utils.INFER and infer_no_label:  # label \t sentence1 \t sentence2
-    return (text_left, text_right), []
-  return (text_left, text_right), label
-
-
-def load_seq2seq_raw_data(paths):
-  """Load raw data for sequence to sequence to sequence."""
-  text = []
-  for path in paths:
-    with open(path, 'r', encoding='utf-8') as file_input:
-      lines = file_input.readlines()
-      for _, line in enumerate(lines):
-        line = line.strip()
-        text.append(line)
-      all_lines = len(lines)
-      del lines
-    logging.info("Data loaded from {}. " \
-                 "Total {} lines".format(path, all_lines))
-  return text
-
-
 def save_a_text_cls_file(label, texts_after, new_path, no_label):
   """Save a text classification data to a file."""
   logging.info("Saving processed file to: {}".format(new_path))
